+++ src/main/java/shop/dalda/order/application/OrderService.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.springframework.stereotype.Service;
-import shop.dalda.exception.UserNotFoundException;
+import shop.dalda.exception.user.auth.UserNotFoundException;
 import shop.dalda.exception.order.OrderNotBelongToUserException;
 import shop.dalda.exception.template.OrderNotFoundException;
 import shop.dalda.exception.template.TemplateNotFoundException;
